refactor: move mobile web auth adapter under core adapters
Implementation and tests now live in src/core/adapters alongside the
other adapters; the accounts/mobileWebAuth subpaths re-export from
there. tempoWallet now requires an explicit baseUrl, and the adapter no
longer sends session meta: the wallet identifies the app through the
consumer discovery document served at baseUrl.
